builder: use fully qualified classnames in XML files
In Checkstyle 7.8 a new mechanisim is used to lookup rules by
name. This new method causes an exception to be thrown is there are
more than one classes found that match the rule name.

The solution is to use fully qualified class names in the module tag
in the XML ruleset files.